admin-orders: separate components
