+++ frontend/src/components/index.js
@@ -2,5 +2,7 @@ export { default as Footer } from './Footer';
 export { default as Header } from './Header';
 export { default as PopularItems } from './PopularItems';
 export { default as Toast } from './Toast';
+export { default as AdminOrderDetails } from './admin/AdminOrderDetails';
+export { default as AdminOrdersIndex } from './admin/AdminOrdersIndex';
 export { default as AdminProductForm } from './admin/AdminProductForm';
 export { default as AdminProductsIndex } from './admin/AdminProductsIndex';
